Fix error when opening new files containing quotes
`vim-fetch` gave this error when opening a new file that contained
quotes:

    Error detected while processing VimEnter Autocommands for "*":
    E116: Invalid arguments for function fetch#buffer

For example, `vim '"ok".txt'` would give an error if `"ok".txt` didn't
exist.

This was happening because the quotes weren't being properly escaped.
This fixes that and removes the error.

I tested this with my personal configuration and on a completely fresh
virtual machine.